pyproject.toml
[project]
name = "game-wizard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.38.3",
    "fastapi[standard]>=0.115.12",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.9.1",
    "sqlmodel>=0.0.24",
    "unicorn>=2.1.3",
]
